👌 IMPROVE: Ajout de test sur mon programme et renommaege du fichier exo2
